Split array.go into multiple go files
This change makes the code easier to read and maintain.

This commit splits array.go into:
- array_data_slab.go
- array_data_slab_decode.go
- array_data_slab_encode.go
- array_extradata.go
- array_iterator.go
- array_metadata_slab.go
- array_metadata_slab_decode.go
- array_metadata_slab_encode.go
- array_size_consts.go
- array_slab.go
- buffer.go

While at it, this commit also groups related functions together.
